% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetLogic.R, R/ConceptSetLogic_webapi.R
\name{.authWebApi}
\alias{.authWebApi}
\title{Cross platform auth for WebApi}
\usage{
.authWebApi(webApiUrl)

.authWebApi(webApiUrl)
}
\description{
Cross platform auth for WebApi

Cross platform auth for WebApi
}
