% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetLogic_webapi.R
\name{.phenelopeBinDirs}
\alias{.phenelopeBinDirs}
\title{Fixed clinical-category bin sub-directory names under a disease folder}
\usage{
.phenelopeBinDirs
}
\description{
Fixed clinical-category bin sub-directory names under a disease folder
}
