% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaprFunctions.R
\name{greateDemographicCriteria}
\alias{greateDemographicCriteria}
\title{Create gender criteria for templates}
\usage{
greateDemographicCriteria(
  minAge = NULL,
  maxAge = NULL,
  male = FALSE,
  female = FALSE
)
}
\arguments{
\item{minAge}{minumum age to apply at index (NULL default)}

\item{maxAge}{maximum age to apply at index (NULL default)}

\item{male}{require male gender record?}

\item{female}{require female gender record?}
}
\description{
Create gender criteria for templates
}
