% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetLogic.R, R/ConceptSetLogic_webapi.R
\name{mergeConceptSets}
\alias{mergeConceptSets}
\title{Merge concept sets}
\usage{
mergeConceptSets(..., webApiUrl, vocabularySourceKey = NULL)

mergeConceptSets(..., webApiUrl, vocabularySourceKey = NULL)
}
\arguments{
\item{...}{concept sets}
}
\description{
Merge multiple concept sets together, resolve them firsts

Merge multiple concept sets together, resolve them firsts
}
