% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetLogic.R
\name{resolveConceptSet}
\alias{resolveConceptSet}
\title{Resolve a concept set so it only includes individual concepts}
\usage{
resolveConceptSet(cs, connection, vocabularyDatabaseSchema)
}
\arguments{
\item{cs}{A concept set expression (list of items with concept and flags)}

\item{connection}{A DatabaseConnector connection object}

\item{vocabularyDatabaseSchema}{The schema containing the OMOP vocabulary tables}
}
\value{
A data frame of resolved concept IDs
}
\description{
Resolve a concept set so it only includes individual concepts
}
