% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaprFunctions.R
\name{resolveHospitalizationVariants}
\alias{resolveHospitalizationVariants}
\title{Resolve which requiresHospitalization variants to emit for a care setting}
\usage{
resolveHospitalizationVariants(expectedCareSetting)
}
\arguments{
\item{expectedCareSetting}{One of `"acute_care_expected"`,
`"acute_care_common"`, `"outpatient_expected"`.}
}
\value{
A logical vector of `requiresHospitalization` values to cross
  (length 1 for expected/outpatient, length 2 for common).
}
\description{
Resolve which requiresHospitalization variants to emit for a care setting
}
